+++ src/refactoring/hooks/admin/index.ts
@@ -1,1 +1,2 @@
 export * from "./useProductForm";
+export * from "./useProductUpdateForm";
